IPOPT/requirements.txt
Django
python-memcached
djangotoolbox
mongoengine
django-mongonaut
django-gravatar2
